+++ src/data/pagesdata/global/home.json
@@ -342,7 +342,6 @@
           "icon": "/img/benefits/easy-sync.svg",
           "text": "Easy Synchronization with Tally"
         },
-
         {
           "icon": "/img/benefits/built-in-white-label.svg",
           "text": "Built-In White Label Option"
