+++ src/data/pagesdata/in/home.json
@@ -351,7 +351,6 @@
           "icon": "/img/benefits/easy-sync.svg",
           "text": "Easy Synchronization with Tally"
         },
-
         {
           "icon": "/img/benefits/built-in-white-label.svg",
           "text": "Built-In White Label Option"
