+++ src/pages/_app.js
@@ -7,6 +7,8 @@ import GlobalComponents from "@/components/globalComponents";
 import { useEffect } from "react";
 import { useRouter } from "next/router";
 import Toastify from "@/components/toastify";
+import getPageData from "@/utils/getPageData";
+import getPageInfo from "@/utils/getPageInfo";
 
 export default function MyApp({ Component, pageProps }) {
   const router = useRouter();
@@ -56,13 +58,18 @@ export default function MyApp({ Component, pageProps }) {
   useEffect(() => {
     require("bootstrap/dist/js/bootstrap.bundle.min.js");
   }, []);
+
+  const rawPath = router.asPath?.split("#")[0]?.split("?")[0];
+  const pageInfo = getPageInfo(rawPath);
+  const pageData = getPageData(pageInfo);
+
   return (
     <>
       {loginSignupPathStatus ? (
         <Navbar browserPath={rawBrowserPath} path={path} />
       ) : null}
       <Header browserPath={rawBrowserPath} path={path} />
-      <Component path={path} {...pageProps} />
+      <Component path={path} {...pageProps} pageData={pageData} pageInfo={pageInfo} />
       {loginSignupPathStatus ? <Footer path={path} /> : null}
       <GlobalComponents path={arrayBrawserPath} />
       <Toastify />
